LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/closed-attribute.window-expected.txt


Harness Error (TIMEOUT), message = null

PASS closed and same-origin nested browsing context 
FAIL closed/close() and same-origin auxiliary browsing context assert_equals: expected true but got false
TIMEOUT closed and cross-origin nested browsing context Test timed out
TIMEOUT closed/close() and cross-origin auxiliary browsing context Test timed out
TIMEOUT closed and cross-site nested browsing context Test timed out
TIMEOUT closed/close() and cross-site auxiliary browsing context Test timed out
